refactor database facade
