+++ plep_main/src/deltadak/ui/AbstractController.java
@@ -1,9 +1,11 @@
 package deltadak.ui;
 
-import deltadak.Database;
 import deltadak.HomeworkTask;
 import deltadak.commands.DeleteCommand;
-import javafx.scene.control.TreeView;
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.GridPane;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -12,14 +14,8 @@ import java.util.List;
  * Abstract controller, useful for testing {@link DeleteCommand} and such things, because then you can provide a dummy
  * controller to the {@link DeleteCommand} while testing.
  */
+@SuppressWarnings("ALL")
 public interface AbstractController {
-
-    /**
-     * Updates database using the given homework tasks for a day.
-     * @param day Date from which the tasks are.
-     * @param homeworkTasks Tasks to be put in the database.
-     */
-    void updateDatabase(LocalDate day, List<List<HomeworkTask>> homeworkTasks);
 
     /**
      * removes empty rows, and then fills up with empty rows
@@ -27,5 +23,10 @@ public interface AbstractController {
      * @param list to clean up
      */
     void cleanUp(TreeView<HomeworkTask> list);
+
+    /**
+     * Getters for the fxml references.
+     */
+    ProgressIndicator getProgressIndicator();
 
 }
